BackEnd/requirements.txt
Click==7.0
Django==3.0.7
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SocketIO==4.2.1
httplib2==0.18.0
idna==2.8
image==1.5.28
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
python-dateutil==2.8.1
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
urllib3==1.25.8
Werkzeug==1.0.0
stream-chat
